split up SignUp{Form,Page} into 2 separate files, 1 per component
This is good practice anyway and makes it easier to lazily-load them using Webpack code splitting using a technique I'm working on. Even without that, this change is worthwhile.
